+++ main.go
@@ -3,7 +3,6 @@ package main
 import (
 	"fmt"
 	"log"
-	"zero-provers/server/config"
 	"zero-provers/server/grpc"
 	"zero-provers/server/http"
 	"zero-provers/server/modes"
@@ -12,8 +11,36 @@ import (
 	"github.com/spf13/cobra"
 )
 
+type Config struct {
+	// Port of the gRPC server.
+	GRPCServerPort int
+	// Port of the HTTP server.
+	HTTPServerPort int
+	// URL path of the HTTP server save endpoint.
+	HTTPServerSaveEndpoint string
+
+	// Directory in which proofs are stored.
+	ProofsOutputDir string
+	// Directory in which mock data is provided.
+	MockDataDir        string
+	MockDataBlockFile  string
+	MockDataStatusFile string
+	MockDataTraceFile  string
+
+	// Mode of the mock server, either static or dynamic.
+	// - static: the server always return the same mock block data.
+	// - dynamic: the server returns new mock block data every x requests.
+	// - random: the server returns random block data every requests.
+	Mode string
+
+	// Set to true if debug mode is enabled.
+	Debug bool
+	// Verbosity of the logs.
+	LogLevel zerolog.Level
+}
+
 func main() {
-	var config config.Config
+	var config Config
 	var rootCmd = &cobra.Command{
 		Use:   "edge-grpc-mock-server",
 		Short: "Edge gRPC mock server",
